[#122] refactor: sort 로직 함수화
